add more session reconnect attempts + unify config setup for magpie/adapter + add retry attempt if session/transaction error still happens
