+++ config/magpie.ini
@@ -89,6 +89,16 @@ cache.acl.enabled = false
 cache.service.enabled = false
 # cache.service.expire = 10
 
+# amount of retries to allow following failed requests for specific
+# database session/transaction errors (usually related to cache handling/reset timing)
+retry.attempts = 2
+
+# output request authentication details in logs
+# WARNING:
+#   Enabling this feature will leak important authentication details in debug logs.
+#   Magpie logger in below section must also be set to DEBUG level.
+magpie.debug_cookie_identity = false
+
 [app:api_app]
 use = egg:Paste#static
 document_root = %(here)s/ui/swagger
